Handle radosgw hosts placement with non-default cluster name
In cephadm-adopt.yml TASK "Update the placement of radosgw hosts"
does not handle when Ansible var cluster is something other than
"ceph", unless this patch is used.

If cluster is set to something other than "ceph", then pass
use extra_container_args to mount the admin keyring with a
non-default name into the container with the default name.
If cluster has the default value, or is unset, then the var
extra_container_args is unchanged.

This patch is necessary to migrate older clusters which were
deployed when custom names were supported.

Closes: https://issues.redhat.com/browse/RHCEPH-10442

Signed-off-by: John Fulton <fulton@redhat.com>